Extract Item.Delete and Item.Destroy
